issue 페이지 별도 모듈로 분리
